requirements.txt
asgiref==3.8.1
autopep8==2.3.2
Brotli==1.1.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.6
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
h11==0.14.0
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.5
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
whitenoise==6.9.0
